decls with aliases now print (without duplications) as   sort {A,B,C} ...   op {x,y,z} ...
For example, given the sort above, there are three keys in
the sort map to the same SortInfo, but only the access via
"A" will print that information.

translate of such structures still needs work, though,
since it seems to translate each sort/op map entry,
using the key as the new target.  That garbles things.


git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@190 7b97033b-253e-4a5c-9b07-e25f9089a9fd