Fix for bug 23: Colimits and spec unions of multiply-named sorts/ops can get confused.
If multiple specs are combined (e.g. via imports or colimit) and
they refer to sorts or ops with overlapping sets of names (e.g.
{A,B,C} and {A}), the resulting spec may erroneously duplicate
the problematic sort/op, e.g.:

  sort {A,B,C}
  sort {C,B,A}

This may cause confusion among alternative references to the problematic
sort/op, e.g. it might cause complaints that the following are duplicate
declarations:

  op f : A
  op f : C

I don't have a small test case yet, but processing all of flaws will
trigger it in the final colimit.

The change to SpecUnion is crucial.
The change to Utilities is a little for performance and more for a better error msg.


git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@650 7b97033b-253e-4a5c-9b07-e25f9089a9fd