Fixed minor bug with special tokens such as ".."
Code that prevented "option" from being parsed as "op" "tion"
 or "deflate" from being parsed as "def" "late", etc.
 was preventing "A..B" from being parsed as "A" ".." "B",
 so it became  "A"  "."  "."  "B" instead.

Now if ".." is declared as a special token, then "..+"
(where + is any non-alphanumeric character) will be kept
intact, but "..A" (where A is any alphabetic character)
will break as ".." "A", etc.  The important feature is
that ".." will not be factored into two dots.

Similarly, if "op" is a special token, "op++" breaks as
"op" "++", but "option" will remain intact.  [This is
less problematic since there is no tension to break
"op" into "o" "p".]


git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@1403 7b97033b-253e-4a5c-9b07-e25f9089a9fd