Added hook to get shadowing directories:  i.e., dirs can be placed "in front of" a base directory, so that all UnitIds (both UnitId_Relative and SpecPath_Relative references) are sought in the shadowing dirs before being sought in the base dir.
First an alias, if any, is computed, so that /tmp files can
be viewed as if they were on the connected dir for commands
like :sw   [This feature was recently added by Stephen.]

Then the possible shadowing dirs are computed using
SpecCalc.shadowingPaths.  (A full list of shadowing dirs
is created for any file under any dir in any such list.)

Then a list of potential files (one per shadowing dir)
is generated.  The first existing file on that list is used.

If SpecCalc.shadowingPaths is null, the behavior should
be unchanged from previous version.


git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@2212 7b97033b-253e-4a5c-9b07-e25f9089a9fd