fixed maddeningly obscure bug:
If an op (e.g. List.map) was polymorphic and had a def but no decl,
addUnCurriedOps was completely losing the type information
on the old op (List.map), so even though it was corretly inserting
the new op (List.map-1-1), people looking for the old op (List.map)
were getting confused by it's bogus type.
Now if there was no decl, one is inferred from the def.


git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@2345 7b97033b-253e-4a5c-9b07-e25f9089a9fd