In case statements, we replace a dispatch on cexpr0 with an assignment of var to cexpr0 followed by a dispatch on var.
In this situation, we can anticipate that we will be doing struct
accesses on cexpr0, but in general, cexpr0 may be too complex to
appear in a C struct accessor form such as (cexpr0 -> attr), so
we need to replace such forms by (var -> attr).

And as long as we're at it, we might just as well replace all the
cexpr0 occurrances by var, not just those appearing in struct
accessors.  (Yell at jlm if this latter assumption is faulty.)


git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@2865 7b97033b-253e-4a5c-9b07-e25f9089a9fd