Revise code for merging OpInfos and SortInfos to use equivSort? and equivTerm? instead of equalSort? and equalTerm?
This required establishing a context spec in which the equivalence
would be done.  Currently that's the spec in progress, but perhaps
it could something more limited such as the base spec.

A side effect of this change is that more common code was pushed
down into MergeSpecs.sw from the four specs that use it.


git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@3212 7b97033b-253e-4a5c-9b07-e25f9089a9fd