----------------------------------------------------------------------
Committing in revised definition of "finite?" for easier proofs of
Isabelle obligations.  This makes the definition use tight prefixes of
the naturals which map exactly onto the set, rather than a loose prefix
which may also map onto other elements not in the set, as long as they
do completely cover the set.

Modified Files:
Sets.sw
----------------------------------------------------------------------


git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@5243 7b97033b-253e-4a5c-9b07-e25f9089a9fd