----------------------------------------------------------------------
Committing in modified version of Sets spec, including proofs and
attempts at proofs.  These are partial or failed for many proofs that
deal with the difficult topics of finite sets or the size of sets.
The ergonomics of the current tools is shown in these examples, to be
not user-friendly.  Even more difficult is probably the fold function,
which has not yet been attempted.  However, this is a first step to
try to prove the realization of this spec in Isabelle.

Modified Files:
Sets.sw
----------------------------------------------------------------------


git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@5245 7b97033b-253e-4a5c-9b07-e25f9089a9fd