pervasive internal cleanup affecting about 350 files
Among other things:

'sort' should now be fully removed in favor of 'type'
(but 'sort' is still accepted for the time being)
As a side effect, some files that had "Sort" in their names
have been renamed to use "Type", to avoid confusion.

The type 'Boolean' should now be fully removed in favor of 'Bool'
(but 'Boolean' is still defined as 'Bool' for now)

The internal types for structures in the Metaslang abstract syntax
are now universally referenced as 'MSType', 'MSTerm', and 'MSPattern',
to make it easier to avoid conflicts with descriptions of types, terms,
and patterns in other langauges such as SpecCalculus, Lisp, Java, C,
Haskell, Isabelle, etc.

Some imports have been restructured to more cleanly share declarations.

And probably some other minor changes as well...


git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@7005 7b97033b-253e-4a5c-9b07-e25f9089a9fd