Mergerules bugfix for not-fully-defined terms.
Previously, mergeRules would insert if branches on terms, even if the
terms happened to reference a existentially-bound variable (e.g. 'x')
that had not been 'defined' with an equation (x = t), where t is doesn't
reference existentially-bound variables.

This behavior is fixed. However, the way it's implemented currently will
cause mergeRules to loop if a clause in the DNF doesn't contain any
conjuncts that either are fully defined or that are a definition.

git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@8249 7b97033b-253e-4a5c-9b07-e25f9089a9fd