Major update to MergeRules.
This should eliminate many of the problems and limitations from before.

There's some more debugging output, which can make it easier to
understand why something is failing.

The transform now correctly supports definitions of the form

  (h',res) = <expr>

This only works, however, if h' and res are local variables. If they are
next-state variables or outputs, the transform will not recognize them
correctly.

git-svn-id: svn+ssh://svn.kestrel.edu/Specware/trunk@8274 7b97033b-253e-4a5c-9b07-e25f9089a9fd